LPS-118855 Merge TypedProperties and Properties
